The constructor should take a string
I'm using named parameters to dictate the names of the parameter
(evidently) in the generated code. I'm trying to do london style TDD, so
I should only do the minimal code required to make the test pass. If I
have some idea of what to do next, I should express it as tests first.

This isn't how I think this constructor should be either. The Client
should not be responsible of parsing the connection string.