The constructor should throw when I pass null
Good, a failing test. I could have started with passing null in the
first test, but then I wouldn't have been set the type of the parameter
of the constructor.

Perhaps starting with a constructor that takes object and then refining
the type (either via static analysis or TDD) sounds like a good idea.
But then when I pass a string it will pass.

This is not a dynamic language where I could write tests that try to
pass things that are not correct, expect exceptions and have those tests
pass. I whould be setting the static type I need, but there I can't
write those tests because they are compilation errors.

This suggest that there are aspects of TDD (the methodology, not the
idea of unit testing) that are better suited for dynamic languages.
Another reason to try to do it this way.

---

As I said, I already know the constructor should not be like this. It
should be parsing the string. I'm playing the game.