I'm playing dumb
This is valid code to make both tests pass.

Arguebly it is as short as the correct code. So, there is no reason to
have selected the correct code over this one if I follow the methodlogy
blindly.

It is obvious in this case: I should thow ArgumentNullException when the
argument is null, and only when the argument is null. But it may not be
obvious in more complex code.

How can we prevent this kind of mistakes in a systematic way?

What test could I add to make this right?

---

I could not possible test all possible string values to see that they
don't throw. Although adding enough would make the correct code be the
minimal code that works (because in if I insist in testing the way I do
here, I would have to check for all the inputs of all the tests).

This is an argument for adding DDT (Data Driven Testing) to TDD (Test
Driven Design).

But, but, but. There is nothing indicating that DDT is the right
solution. It is just that I know that it is a solution.