Added Metaphor project
My intent with this project is to provide a way to generate strings.

I see the need to inject another process into TDD, one that would allow
me to select a third party library.

But, TDD suggests that if the code is not mine, I should be mocking it,
and if it is mine I should be testing it. TDD seems to encourage
Not-Invented-Here syndrome.

---

If I am creating project A by client requirements, and I decide I want
to create project B that cover some need of project A... project B will
be what is strictly needed for project B, unless I insert my own
requirements.