Named it ItemPicker and make it generic
I have been thinking about TDD and DDT and how to decide what tests to
do...

I will not follow dogma and let TDD and unit testing be the only "true"
of testing. For me DDT makes more sense, yet not alone. Threading and
performance problems will never be found without higher level testing.

With that said, I have decided that I must create test for every
*output" or verifiable state. And I must write these tests as
documentation. The tests must allow the reader to know what are the
*ranges of values* that lead to each *output*.

---

While trying to figure out how to do that for my CharacterPicker, I
found out that a char is not a character.

*What?*

.NET is not better that PHP! you can't really change the internal
encoding of strings! - ok, that is an exageration, but really you can't,
all strings are UTF-16. And a char is a UTF-16 unit, not a code point,
meaning that a code point sometimes is two char.

Thus, if I want to pick characters, I need to use string.

Well, screw that, make the class generic! - if you want string, you get
string, if you want char, you get char. For the me... I will be testing
with int.